requirements.txt
openai
pymupdf
apify-client
python-dotenv
streamlit

-e .